Honeypots (#23)
* [feat]: Honeypot manager init

Created the honeypot manager with some basic features

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [feat]: Honeypot Manager

Newly updated Honeypot Manager.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [feat]: Honeypot Manager

Newly updated Honeypot Manager.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Return type to the methods

I added the proper return methods to indicate what they will be returning.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Update Honeypot Manager

Update position of the files for the honeypot manager. Place them in their own folder structure to make it cleaner.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Update Structure

I created a new app folder for the backend stuff, and moved the entire code structure to that. I will be following the MVC architecture from here onwards. The folder structure has been made that way.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Update API Code

Update the API code to work properly and removed most of the complexity.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Update API Code

Update the API code to work properly and removed most of the complexity.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [refactor]: Update Entire Container Management

The path will be taken to go for a form of Microservice architecture. Since it will help the project be modular and expandable. The entire backend honeypot manager has been refactored.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [fix]: Update Error Catch

Updated code to show the proper error when podman isn't running.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>

* [chore]: Minor Changes (#12) (#15)

Just updated the .gitignore and a new init

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* Ftp honeypot (#17)

* [chore]: Minor Changes (#12) (#13)

Just updated the .gitignore and a new init

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [feat]: FTP Honeypot with other update

The following are the changes,

    - Created the FTP Folders.
    - Created the FTP honeypot, created a FTP server and then converted that to the honeypot.
    - Update the Honeypot Class to dynamically check for the available honeypots.
    - Update exception to tell the user if a honeypot exists or not.
    - New honeypot_configs.yaml to add all of the honeypots that exist in the system.

Breaking changes,
    - honeypot_manager_controller has new updates.
    - exceptions have new updates.
    - Honeypot class has new updates.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

---------

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* Ftp honeypot (#22)

* [chore]: Minor Changes (#12) (#13)

Just updated the .gitignore and a new init

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [feat]: FTP Honeypot with other update

The following are the changes,

    - Created the FTP Folders.
    - Created the FTP honeypot, created a FTP server and then converted that to the honeypot.
    - Update the Honeypot Class to dynamically check for the available honeypots.
    - Update exception to tell the user if a honeypot exists or not.
    - New honeypot_configs.yaml to add all of the honeypots that exist in the system.

Breaking changes,
    - honeypot_manager_controller has new updates.
    - exceptions have new updates.
    - Honeypot class has new updates.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [feat]: FTP Automatic File Creation

I added the ability to automatically create files in the FTP honeypot.
I also fixed a bug where the honeypot would not be created because it couldn't find the path to the honeypot.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

---------

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

---------

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>
Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>