[feat]: Pub/Sub with NATS
Implemented the base machanism for the honeypots to send their logs.
Current implementation only works with the SSH Honeypot.
The SSH Honeypot will connect to the NATS server running on the local machine, then make a topic if no new one exisits.
Then when an attacker is connected successfully with the right authenication, the data will start getting collected.
Then when the session ends or get terminated, the data will be published to the server and where the subscriber can get that data.
For now we are only prininting the logs.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>