Logger machanics (#37)
* [feat]: Pub/Sub with NATS

Implemented the base machanism for the honeypots to send their logs.
Current implementation only works with the SSH Honeypot.
The SSH Honeypot will connect to the NATS server running on the local machine, then make a topic if no new one exisits.
Then when an attacker is connected successfully with the right authenication, the data will start getting collected.
Then when the session ends or get terminated, the data will be published to the server and where the subscriber can get that data.
For now we are only prininting the logs.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [fix]: Updated Log Collector with NATS

Fixed a bit of the old Log Manager system,

- Added the ability to create the NATS Server through the code propperly.
- Added the ability to create teh log collector through the code proplery.
- A FastAPI service is in place, to access the NATS server and the log collector.
- The log collector and NATS Server both have the ability to connect to a podman internal network. This make the pub sub more private since it's not
exposed to the external network or even binded to it.

This way the NATS Server won't be accessible from outside. Meaning no one expect for anything within the podman network created will be able to connect.

Fixes #36

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

* [feat]: Inter Container Communication

The NATS server, log collector, and Honeypots are connected to the same network and can communicate with each other.
The Honeypot (SSH only) has the publisher, that connects to the NATS Server and Stream and sends the data to the log collectory properly.

Breaking Changes,
- The SSH Server class has been stripped down due to testing purposes.
- The output ins't good on the SSH Honeypot.
- Honeypot manager has been updated to send the NATS server url through enviorment variables.

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>

---------

Signed-off-by: Shaveen-Balasooriya <shaveenb2004@outlook.com>