[feat]: Working Log Manager
I have created the log manager to properly start up the service need such as the NATS server, OpenSearch server, OpenSearch dashboard, and the log collector. The log manager is responsible for managing the log collection process, including starting and stopping the log collector, configuring the NATS server, and managing the OpenSearch server.
The log manger is now fixed to using the SubProcesses library entirely without the need for the Podman SDK.
Much more cleaner and easier to understand.

Breaking changes,

honeypot Manager - This needs to be refactored as well to use the SubProcesses instead of the Podman SDK.

Signed-off-by: Shaveen Balasooriya <shaveenb2004@outlook.com>